cantidaddes en disponibilidad float

+++ back/src/db/migrations/20240213110847-create-disponibilidad-articulos.js
@@ -10,13 +10,13 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       cant_fisica: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       cant_comprometida: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       cant_disponible: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       createdAt: {
         allowNull: false,
